Refactor `ConnectionsMenu` into multiple sub-components
