+++ src/widgets/connectionsMenu/connectionsMenu.tsx
@@ -1,38 +1,38 @@
-import cx from 'clsx';
 import * as React from 'react';
+import cx from 'clsx';
 
-import { EventObserver } from '../coreUtils/events';
-import { useTranslation, type Translation, TranslatedText } from '../coreUtils/i18n';
+import { EventObserver } from '../../coreUtils/events';
+import { useTranslation, type Translation, TranslatedText } from '../../coreUtils/i18n';
 
-import { ElementModel, ElementIri, LinkTypeIri, LinkTypeModel } from '../data/model';
-import { generate128BitID, makeCaseInsensitiveFilter } from '../data/utils';
+import type { ElementModel, ElementIri, LinkTypeIri, LinkTypeModel } from '../../data/model';
 
-import { CanvasApi, useCanvas } from '../diagram/canvasApi';
-import { changeLinkTypeVisibility, placeElementsAroundTarget } from '../diagram/commands';
-import { Element, VoidElement } from '../diagram/elements';
-import { getContentFittingBox } from '../diagram/geometry';
-import { DiagramModel } from '../diagram/model';
-import { RenderingLayer } from '../diagram/renderingState';
-import { HtmlSpinner } from '../diagram/spinner';
+import { CanvasApi, useCanvas } from '../../diagram/canvasApi';
+import { changeLinkTypeVisibility, placeElementsAroundTarget } from '../../diagram/commands';
+import { Element, VoidElement } from '../../diagram/elements';
+import { getContentFittingBox } from '../../diagram/geometry';
+import { DiagramModel } from '../../diagram/model';
+import { RenderingLayer } from '../../diagram/renderingState';
 
-import { BuiltinDialogType } from '../editor/builtinDialogType';
+import { BuiltinDialogType } from '../../editor/builtinDialogType';
+import { requestElementData, restoreLinksBetweenElements } from '../../editor/dataDiagramModel';
+import { EntityElement, EntityGroup, iterateEntitiesOf } from '../../editor/dataElements';
+
+import { SearchInput, SearchInputStore, useSearchInputStore } from '../utility/searchInput';
+
+import type { InstancesSearchCommands } from '../instancesSearch';
+
+import { ConnectionsMenuTopic, InstancesSearchTopic } from '../../workspace/commandBusTopic';
 import {
-    requestElementData, restoreLinksBetweenElements, getAllPresentEntities,
-} from '../editor/dataDiagramModel';
-import { EntityElement, EntityGroup, iterateEntitiesOf } from '../editor/dataElements';
-import { WithFetchStatus } from '../editor/withFetchStatus';
+    type WorkspaceContext, WorkspaceEventKey, useWorkspace,
+} from '../../workspace/workspaceContext';
 
-import { SearchInput, SearchInputStore, useSearchInputStore } from '../widgets/utility/searchInput';
-import type { InstancesSearchCommands } from '../widgets/instancesSearch';
-
-import { type WorkspaceContext, WorkspaceEventKey, useWorkspace } from '../workspace/workspaceContext';
-import { ConnectionsMenuTopic, InstancesSearchTopic } from '../workspace/commandBusTopic';
-
-import { highlightSubstring } from './utility/listElementView';
-import { SearchResults } from './utility/searchResults';
 import {
-    TreeList, type TreeListModel, type TreeListRenderItem, type TreeListFocusProps,
-} from './utility/treeList';
+    PropertySuggestionHandler, PropertyScore, SortMode, ConnectionsData, ConnectionSuggestions,
+    ConnectionCount, ObjectsData, LinkDataChunk, ObjectPlacingMode,
+    CLASS_NAME, LINK_COUNT_PER_PAGE, LoadingSpinner,
+} from './menuCommon';
+import { ConnectionsList } from './connectionList';
+import { EntityList } from './entityList';
 
 /**
  * Props for {@link ConnectionsMenu} component.
@@ -83,57 +83,6 @@ export interface ConnectionsMenuCommands {
          */
         readonly openAll?: boolean;
     };
-}
-
-/**
- * Provides smart suggestions when searching by the link type label.
- *
- * @see {@link ConnectionsMenuProps.suggestProperties}
- */
-export type PropertySuggestionHandler = (params: PropertySuggestionParams) => Promise<PropertyScore[]>;
-
-/**
- * Parameters for the smart link type suggestion handler.
- *
- * @see {@link PropertySuggestionHandler}
- */
-export interface PropertySuggestionParams {
-    /**
-     * Target connected entity IRI.
-     */
-    elementId: string;
-    /**
-     * Link type label search token.
-     */
-    token: string;
-    /**
-     * A collection of possible link type IRIs.
-     */
-    properties: readonly string[];
-    /**
-     * Current diagram model data language.
-     */
-    lang: string;
-    /**
-     * Cancellation signal.
-     */
-    signal: AbortSignal | undefined;
-}
-
-/**
- * Result entry for the smart link type suggestion handler.
- *
- * @see {@link PropertySuggestionHandler}
- */
-export interface PropertyScore {
-    /**
-     * Link type IRI.
-     */
-    propertyIri: string;
-    /**
-     * Suggestion score (higher is more suggested for the top positions).
-     */
-    score: number;
 }
 
 /**
@@ -300,39 +249,6 @@ interface MenuState {
     readonly connectionSuggestions: ConnectionSuggestions;
     readonly objects?: ObjectsData;
 }
-
-type SortMode = 'alphabet' | 'smart';
-
-interface ConnectionsData {
-    readonly links: ReadonlyArray<LinkTypeModel>;
-    readonly counts: ReadonlyMap<LinkTypeIri, ConnectionCount>;
-}
-
-interface ConnectionCount {
-    readonly inexact: boolean;
-    readonly inCount: number;
-    readonly outCount: number;
-}
-
-interface ObjectsData {
-    readonly chunk: LinkDataChunk;
-    readonly elements: ReadonlyArray<ElementModel>;
-}
-
-interface LinkDataChunk {
-    /**
-     * Random key to check if chunk is different from another
-     * (i.e. should be re-rendered).
-     */
-    readonly chunkId: string;
-    readonly linkType: LinkTypeModel;
-    readonly direction?: 'in' | 'out';
-    readonly expectedCount: number | 'some';
-    readonly pageCount: number;
-}
-
-const CLASS_NAME = 'reactodia-connections-menu';
-const LINK_COUNT_PER_PAGE = 100;
 
 class ConnectionsMenuInner extends React.Component<ConnectionsMenuInnerProps, MenuState> {
     private readonly ALL_RELATED_ELEMENTS_LINK: LinkTypeModel;
@@ -648,7 +564,7 @@ class ConnectionsMenuInner extends React.Component<ConnectionsMenuInnerProps, Me
 
     private getBody() {
         const {
-            targetIris, connectionSearch, objectSearch, workspace, translation: t,
+            targetIris, connectionSearch, objectSearch, workspace: {getCommandBus},
         } = this.props;
         const {
             panel, connectionSortMode, loadingState, objects, connections, connectionSuggestions,
@@ -658,14 +574,12 @@ class ConnectionsMenuInner extends React.Component<ConnectionsMenuInnerProps, Me
             return <LoadingSpinner error={true} />;
         } else if (objects && panel === 'objects') {
             return (
-                <ObjectsPanel
+                <EntityList key={objects.chunk.chunkId}
                     data={objects}
                     onMoveToFilter={this.onMoveToFilter}
                     filterKey={objectSearch.value}
-                    loading={loadingState === 'loading'}
+                    isLoading={loadingState === 'loading'}
                     onPressAddSelected={this.onAddSelectedElements}
-                    workspace={workspace}
-                    translation={t}
                 />
             );
         } else if (connections && panel === 'connections') {
@@ -678,7 +592,7 @@ class ConnectionsMenuInner extends React.Component<ConnectionsMenuInnerProps, Me
                 return <LoadingSpinner />;
             }
 
-            const commands = workspace.getCommandBus(InstancesSearchTopic);
+            const commands = getCommandBus(InstancesSearchTopic);
             const event: InstancesSearchCommands['findCapabilities'] = {capabilities: []};
             commands.trigger('findCapabilities', event);
 
@@ -848,527 +762,4 @@ class ConnectionsMenuInner extends React.Component<ConnectionsMenuInnerProps, Me
             );
         }
     };
-}
-
-interface ConnectionSuggestions {
-    readonly filterKey: string | null;
-    readonly scores: ReadonlyMap<LinkTypeIri, PropertyScore>;
-}
-
-function ConnectionsList(props: {
-    data: ConnectionsData;
-    filterKey: string;
-    sortMode: SortMode;
-    suggestions: ConnectionSuggestions;
-
-    allRelatedLink: LinkTypeModel;
-    onExpandLink: (chunk: LinkDataChunk) => void;
-    onMoveToFilter: ((chunk: LinkDataChunk) => void) | undefined;
-
-    scrolledListRef: React.RefObject<HTMLUListElement | null>;
-}) {
-    const {
-        data, filterKey, sortMode, suggestions,
-        allRelatedLink, onExpandLink, onMoveToFilter, scrolledListRef,
-    } = props;
-    const {model} = useWorkspace();
-    const t = useTranslation();
-
-    const isSmartMode = sortMode === 'smart' && !filterKey;
-
-    const textFilter = filterKey ? makeCaseInsensitiveFilter(filterKey) : undefined;
-    const links = isSmartMode ? [] : (data.links || [])
-        .map(link => model.getLinkType(link.id)?.data ?? link)
-        .filter(link => {
-            const text = t.formatLabel(link.label, link.id, model.language);
-            return !textFilter || textFilter(text);
-        })
-        .sort(makeLinkTypeComparer(t, model.language));
-
-    const probableLinks =  (data.links ?? [])
-        .map(link => model.getLinkType(link.id)?.data ?? link)
-        .filter(link => {
-            const score = suggestions.scores.get(link.id);
-            return !links.includes(link) && score !== undefined && (score.score > 0 || isSmartMode);
-        })
-        .sort(makeLinkTypeComparer(t, model.language, suggestions));
-
-    const regularEntries = getConnectionLinks(links, {
-        counts: data.counts,
-        scores: suggestions.scores,
-    });
-    const probableEntries = getConnectionLinks(probableLinks, {
-        counts: data.counts,
-        scores: suggestions.scores,
-        probable: true,
-    });
-
-    const entries: ConnectionEntry[] = [];
-    if (regularEntries.length > 1 || (isSmartMode && probableEntries.length > 1)) {
-        const countMap = data.counts;
-        const {inCount, outCount, inexact} = countMap.get(allRelatedLink.id)!;
-        const totalCount = inCount + outCount;
-        entries.push({
-            type: 'link',
-            key: allRelatedLink.id,
-            linkType: allRelatedLink,
-            count: inexact && totalCount > 0 ? 'some' : totalCount,
-        });
-        entries.push({type: 'separator'});
-        for (const entry of regularEntries) {
-            entries.push(entry);
-        }
-    }
-
-    if (probableEntries.length > 0) {
-        if (isSmartMode) {
-            entries.push({type: 'probable-hint'});
-            for (const entry of probableEntries) {
-                entries.push(entry);
-            }
-        }
-    }
-
-    const renderItem = React.useCallback<TreeListRenderItem<ConnectionEntry, void>>(
-        ({item, focusProps}) => {
-            if (item.type === 'link') {
-                return (
-                    <LinkInPopupMenu
-                        link={item.linkType}
-                        direction={item.direction}
-                        count={item.count}
-                        filterKey={item.probable ? '' : filterKey}
-                        probability={item.probability}
-                        onExpandLink={onExpandLink}
-                        onMoveToFilter={onMoveToFilter}
-                        focusProps={focusProps}
-                    />
-                );
-            } else if (item.type === 'separator') {
-                return <hr className={`${CLASS_NAME}__links-list-hr`} />;
-            } else if (item.type === 'probable-hint') {
-                return (
-                    <li className={`${CLASS_NAME}__links-probably-label`}>
-                        {t.text('connections_menu.links.suggest_similar')}
-                    </li>
-                );
-            }
-            return null;
-        },
-        [t, filterKey, onExpandLink, onMoveToFilter]
-    );
-
-    const rootProps = React.useMemo((): React.HTMLProps<HTMLUListElement> => ({
-        /* For compatibility with React 19 typings */
-        ref: scrolledListRef as React.RefObject<HTMLUListElement>,
-        className: `${CLASS_NAME}__links-root`,
-        role: 'list',
-    }), []);
-    const forestProps = React.useMemo((): React.HTMLProps<HTMLUListElement> => ({}), []);
-    const itemProps = React.useMemo((): React.HTMLProps<HTMLLIElement> => ({
-        className: `${CLASS_NAME}__links-item`,
-        role: 'listitem',
-    }), []);
-
-    return (
-        <div
-            className={cx(
-                `${CLASS_NAME}__links-list`,
-                'reactodia-scrollable',
-                entries.length === 0 ? `${CLASS_NAME}__links-list-empty` : undefined
-            )}
-            tabIndex={-1}
-        >
-            <TreeList
-                model={ConnectionListModel}
-                items={entries}
-                renderItem={renderItem}
-                rootProps={rootProps}
-                forestProps={forestProps}
-                itemProps={itemProps}
-            />
-            {entries.length === 0 ? (
-                <label className={`${CLASS_NAME}__links-no-results`}>
-                    {t.text('connections_menu.links.no_results')}
-                </label>
-            ) : null}
-        </div>
-    );
-}
-
-const ConnectionListModel: TreeListModel<ConnectionEntry, void> = {
-    getKey: item => item.type === 'link' ? item.key : item.type,
-    getChildren: item => undefined,
-    getDefaultSelected: (item, selected) => undefined,
-    isActive: item => item.type === 'link',
-};
-
-type ConnectionEntry = ConnectionEntryLink | ConnectionEntrySeparator;
-
-interface ConnectionEntrySeparator {
-    readonly type: 'separator' | 'probable-hint';
-}
-
-interface ConnectionEntryLink {
-    readonly type: 'link';
-    readonly key: string;
-    readonly linkType: LinkTypeModel;
-    readonly direction?: 'in' | 'out';
-    readonly count: number | 'some';
-    readonly probable?: boolean;
-    readonly probability?: number;
-}
-
-function getConnectionLinks(links: LinkTypeModel[], options: {
-    counts: ConnectionsData['counts'];
-    scores: ConnectionSuggestions['scores'];
-    probable?: boolean;
-}): ConnectionEntryLink[] {
-    const {counts, scores, probable} = options;
-    const entries: ConnectionEntryLink[] = [];
-    const addView = (link: LinkTypeModel, direction: 'in' | 'out') => {
-        const {inCount, outCount, inexact} = counts.get(link.id) ?? {
-            inCount: 0,
-            outCount: 0,
-            inexact: false,
-        };
-        const count = direction === 'in' ? inCount : outCount;
-        if (count === 0) {
-            return;
-        }
-
-        const postfix = probable ? '-probable' : '';
-        const score = scores.get(link.id);
-        entries.push({
-            type: 'link',
-            key: `${direction}-${link.id}-${postfix}`,
-            linkType: link,
-            direction,
-            count: inexact && count > 0 ? 'some' : count,
-            probable,
-            probability: probable && score !== undefined ? score.score : 0,
-        });
-    };
-
-    for (const link of links) {
-        addView(link, 'in');
-        addView(link, 'out');
-    }
-
-    return entries;
-}
-
-function makeLinkTypeComparer(
-    t: Translation,
-    language: string,
-    suggestions?: ConnectionSuggestions
-): (a: LinkTypeModel, b: LinkTypeModel) => number {
-    return (a, b) => {
-        if (suggestions) {
-            const {scores} = suggestions;
-            const aWeight = scores.has(a.id) ? scores.get(a.id)!.score : 0;
-            const bWeight = scores.has(b.id) ? scores.get(b.id)!.score : 0;
-            if (aWeight > bWeight) {
-                return -1;
-            } else if (aWeight < bWeight) {
-                return 1;
-            }
-        }
-
-        const aText = t.formatLabel(a.label, a.id, language);
-        const bText = t.formatLabel(b.label, b.id, language);
-        return aText.localeCompare(bText);
-    };
-}
-
-function LinkInPopupMenu(props: {
-    link: LinkTypeModel;
-    count: number | 'some';
-    direction?: 'in' | 'out';
-    filterKey?: string;
-    onExpandLink: (linkDataChunk: LinkDataChunk) => void;
-    onMoveToFilter: ((linkDataChunk: LinkDataChunk) => void) | undefined;
-    probability?: number;
-    focusProps?: TreeListFocusProps;
-}) {
-    const {
-        link, filterKey, direction, count, onExpandLink, onMoveToFilter, probability = 0, focusProps,
-    } = props;
-    const {model} = useWorkspace();
-    const t = useTranslation();
-
-    const relation = t.formatLabel(link.label, link.id, model.language);
-    const relationIri = model.locale.formatIri(link.id);
-    const probabilityPercent = Math.round(probability * 100);
-    const textLine = highlightSubstring(
-        relation + (probabilityPercent > 0 ? ` (${probabilityPercent}%)` : ''),
-        filterKey
-    );
-    const title = (
-        direction === 'in' ? t.text('connections_menu.link.source_title', {relation, relationIri}) :
-        direction === 'out' ? t.text('connections_menu.link.target_title', {relation, relationIri}) :
-        t.text('connections_menu.link.both_title', {relation, relationIri})
-    );
-    const navigateTitle = (
-        direction === 'in' ? t.text('connections_menu.link.source_navigate_title', {relation, relationIri}) :
-        direction === 'out' ? t.text('connections_menu.link.target_navigate_title', {relation, relationIri}) :
-        t.text('connections_menu.link.both_navigate_title', {relation, relationIri})
-    );
-
-    const onExpandLinkClick = () => {
-        onExpandLink({
-            chunkId: generate128BitID(),
-            linkType: link,
-            direction,
-            expectedCount: count,
-            pageCount: 1,
-        });
-    };
-
-    const onMoveToFilterClick = (e: React.MouseEvent) => {
-        e.stopPropagation();
-        onMoveToFilter?.({
-            chunkId: generate128BitID(),
-            linkType: link,
-            direction,
-            expectedCount: count,
-            pageCount: 1,
-        });
-    };
-
-    return (
-        <div className={`${CLASS_NAME}__link`}
-            data-linktypeid={link.id}>
-            <button {...focusProps}
-                className={`${CLASS_NAME}__link-button`}
-                title={title}
-                onClick={onExpandLinkClick}
-            >
-                {direction === 'in' || direction === 'out' ? (
-                    <div className={`${CLASS_NAME}__link-direction`}>
-                        {direction === 'in' && <div className={`${CLASS_NAME}__link-direction-in`} />}
-                        {direction === 'out' && <div className={`${CLASS_NAME}__link-direction-out`} />}
-                    </div>
-                ) : null}
-                <WithFetchStatus type='linkType' target={link.id}>
-                    <div className={`${CLASS_NAME}__link-title`}>{textLine}</div>
-                </WithFetchStatus>
-                {count === 'some' ? null : (
-                    <span className={`reactodia-badge ${CLASS_NAME}__link-count`}>
-                        {count <= LINK_COUNT_PER_PAGE ? count : `${LINK_COUNT_PER_PAGE}+`}
-                    </span>
-                )}
-                {onMoveToFilter ? <div className={`${CLASS_NAME}__link-filter-spacer`} /> : null}
-                <div className={`${CLASS_NAME}__link-navigate-button`}
-                    title={navigateTitle}
-                />
-            </button>
-            {onMoveToFilter ? (
-                <button {...focusProps}
-                    className={`${CLASS_NAME}__link-filter-button`}
-                    onClick={onMoveToFilterClick}
-                    title={t.text('connections_menu.link.move_to_filter.title')}
-                />
-            ) : null}
-        </div>
-    );
-}
-
-interface ObjectsPanelProps {
-    data: ObjectsData;
-    loading?: boolean;
-    filterKey?: string;
-    onPressAddSelected: (selected: ElementModel[], mode: ObjectPlacingMode) => void;
-    onMoveToFilter: ((linkDataChunk: LinkDataChunk) => void) | undefined;
-    workspace: WorkspaceContext;
-    translation: Translation;
-}
-
-type ObjectPlacingMode = 'separately' | 'grouped';
-
-interface ObjectsPanelState {
-    chunkId: string;
-    selection: ReadonlySet<ElementIri>;
-}
-
-class ObjectsPanel extends React.Component<ObjectsPanelProps, ObjectsPanelState> {
-    constructor(props: ObjectsPanelProps) {
-        super(props);
-        this.state = ObjectsPanel.makeStateFromProps(props);
-    }
-
-    static getDerivedStateFromProps(
-        props: ObjectsPanelProps,
-        state: ObjectsPanelState | undefined
-    ): ObjectsPanelState | null {
-        if (state && state.chunkId === props.data.chunk.chunkId) {
-            return null;
-        }
-        return ObjectsPanel.makeStateFromProps(props);
-    }
-
-    static makeStateFromProps(props: ObjectsPanelProps): ObjectsPanelState {
-        return {
-            chunkId: props.data.chunk.chunkId,
-            selection: new Set<ElementIri>(),
-        };
-    }
-
-    private getFilteredObjects(): ReadonlyArray<ElementModel> {
-        const {workspace: {model}, data, filterKey} = this.props;
-        if (!filterKey) {
-            return data.elements;
-        }
-        
-        const textFilter = makeCaseInsensitiveFilter(filterKey);
-        return data.elements.filter(element => {
-            const text = model.locale.formatEntityLabel(element, model.language);
-            return textFilter(text);
-        });
-    }
-
-    private updateSelection = (newSelection: ReadonlySet<ElementIri>) => {
-        this.setState({selection: newSelection});
-    };
-
-    private renderCounter(activeObjCount: number) {
-        const {data: {chunk, elements}, translation: t} = this.props;
-        const countString = t.text('connections_menu.entities.counter_label', {
-            count: activeObjCount,
-            total: elements.length,
-        });
-
-        let extraCountInfo: React.ReactElement | null = null;
-        if (chunk.expectedCount !== 'some') {
-            const extraCount = elements.length - Math.min(LINK_COUNT_PER_PAGE, chunk.expectedCount);
-            const extra = Math.abs(extraCount) > LINK_COUNT_PER_PAGE ?
-                `${LINK_COUNT_PER_PAGE}+` : Math.abs(extraCount).toString();
-            extraCountInfo = (
-                <span className={`${CLASS_NAME}__objects-extra`}
-                    title={extraCount === 0 ? undefined : (
-                        extraCount > 0
-                            ? t.text('connections_menu.entities.extra_title', {value: extra})
-                            : t.text('connections_menu.entities.missing_title', {value: extra})
-                    )}>
-                    {extraCount === 0 ? null : (
-                        extraCount > 0
-                            ? t.text('connections_menu.entities.extra_label', {value: extra})
-                            : t.text('connections_menu.entities.missing_label', {value: extra})
-                    )}
-                </span>
-            );
-        }
-
-        return (
-            <div className={`${CLASS_NAME}__objects-count`}>
-                <span>{countString}</span>
-                {extraCountInfo}
-            </div>
-        );
-    }
-
-    render() {
-        const {
-            data, filterKey, onPressAddSelected, onMoveToFilter,
-            workspace: {model},
-            translation: t,
-        } = this.props;
-        const {selection} = this.state;
-        const objects = this.getFilteredObjects();
-
-        const presentEntities = getAllPresentEntities(model);
-        const isAllSelected = objects.every(item =>
-            presentEntities.has(item.id) || selection.has(item.id)
-        );
-
-        const nonPresented = objects.filter(item => !presentEntities.has(item.id));
-        const active = nonPresented.filter(item => selection.has(item.id));
-        const selectedItems = active.length > 0 ? active : nonPresented;
-
-        return <div className={`${CLASS_NAME}__objects`}>
-            <div className={`${CLASS_NAME}__objects-select-all`}>
-                <label>
-                    <input type='checkbox'
-                        name='reactodia-connections-menu-select-all'
-                        checked={isAllSelected && nonPresented.length > 0}
-                        onChange={() => this.updateSelection(new Set<ElementIri>(
-                            isAllSelected ? [] : nonPresented.map(item => item.id)
-                        ))}
-                        disabled={nonPresented.length === 0}
-                        title={t.text('connections_menu.select_all.title')}
-                    />
-                    <span>{t.text('connections_menu.select_all.label')}</span>
-                </label>
-                <div className={`${CLASS_NAME}__objects-spacer`} aria-hidden='true' />
-                {this.renderCounter(active.length)}
-            </div>
-            {this.props.loading ? (
-                <div className={`${CLASS_NAME}__objects-loading`}>
-                    <LoadingSpinner />
-                </div>
-            ) : objects.length === 0 ? (
-                <div className={`${CLASS_NAME}__objects-no-results`}>
-                    {t.text('connections_menu.entities.no_results')}
-                </div>
-            ) : (
-                <div className={`${CLASS_NAME}__objects-list`} tabIndex={-1}>
-                    <SearchResults
-                        items={objects}
-                        selection={this.state.selection}
-                        onSelectionChanged={this.updateSelection}
-                        highlightText={filterKey}
-                    />
-                    {data.chunk.expectedCount !== 'some' && data.chunk.expectedCount > LINK_COUNT_PER_PAGE ? (
-                        onMoveToFilter ? (
-                            <div className={`${CLASS_NAME}__move-to-filter`}
-                                onClick={() => onMoveToFilter(data.chunk)}>
-                                {t.text('connections_menu.entities.truncated_results_expand', {
-                                    limit: LINK_COUNT_PER_PAGE,
-                                })}
-                            </div>
-                        ) : (
-                            <div className={`${CLASS_NAME}__move-to-filter`}>
-                                {t.text('connections_menu.entities.truncated_results', {
-                                    limit: LINK_COUNT_PER_PAGE,
-                                })}
-                            </div>
-                        )
-                    ) : null}
-                </div>
-            )}
-            <div className={`${CLASS_NAME}__objects-statusbar`}>
-                <button
-                    className={cx(
-                        `${CLASS_NAME}__objects-add-button`,
-                        'reactodia-btn reactodia-btn-secondary'
-                    )}
-                    disabled={this.props.loading || selectedItems.length <= 1}
-                    onClick={() => onPressAddSelected(selectedItems, 'grouped')}>
-                    {t.text('connections_menu.entities.add_group')}
-                </button>
-                <button
-                    className={cx(
-                        `${CLASS_NAME}__objects-add-button`,
-                        'reactodia-btn reactodia-btn-primary'
-                    )}
-                    disabled={this.props.loading || nonPresented.length === 0}
-                    onClick={() => onPressAddSelected(selectedItems, 'separately')}>
-                    {active.length > 0
-                        ? t.text('connections_menu.entities.add_selected')
-                        : t.text('connections_menu.entities.add_all')
-                    }
-                </button>
-            </div>
-        </div>;
-    }
-}
-
-function LoadingSpinner(props: { error?: boolean }) {
-    return (
-        <div className={`${CLASS_NAME}__spinner`}>
-            <HtmlSpinner width={30} height={30}
-                errorOccurred={Boolean(props.error)}
-            />
-        </div>
-    );
 }
