+++ src/widgets/connectionsMenu/index.ts
@@ -1,0 +1,6 @@
+export {
+    ConnectionsMenu, type ConnectionsMenuProps, type ConnectionsMenuCommands,
+} from './connectionsMenu';
+export type {
+    PropertySuggestionHandler, PropertySuggestionParams, PropertyScore,
+} from './menuCommon';
